+++ src/types/mock-slot.js
@@ -3,12 +3,14 @@ var cache = require('../modules/slot-cache.js'),
 
 module.exports = function (name) {
     return {
+        gpt: {
+            setTargeting: cache(name).set.targeting
+        },
         mock: true,
         name: name,
         off: stub,
         on: cache(name).set.event,
         one: cache(name).set.single,
-        setTargeting: cache(name).set.targeting,
         trigger: stub
     };
 };